extracting jetty-specific classes
