+++ core/src/main/java/com/glines/socketio/server/transport/XHRPollingTransport.java
@@ -33,13 +33,13 @@ import javax.servlet.http.HttpServletResponse;
 import com.glines.socketio.server.SocketIOFrame;
 import com.glines.socketio.server.SocketIOSession;
 
-public class XHRPollingTransport extends JettyXHRTransport {
+public class XHRPollingTransport extends XHRTransport {
 	public static final String TRANSPORT_NAME = "xhr-polling";
 
-	protected class XHRPollingSessionHelper extends XHRSessionHelper {
+	protected class XHRPollingSessionHelper extends JettyXHRSessionHelper {
 
-		XHRPollingSessionHelper(SocketIOSession session) {
-			super(session, false);
+		XHRPollingSessionHelper(SocketIOSession session, int bufferSize, int maxIdleTime) {
+			super(session, false, bufferSize, maxIdleTime);
 		}
 
 		protected void startSend(HttpServletResponse response) throws IOException {
@@ -72,7 +72,7 @@ public class XHRPollingTransport extends JettyXHRTransport {
 	}
 	
 
-	protected XHRPollingSessionHelper createHelper(SocketIOSession session) {
-		return new XHRPollingSessionHelper(session);
+	protected XHRPollingSessionHelper createHelper(SocketIOSession session, int bufferSize, int maxIdleTime) {
+		return new XHRPollingSessionHelper(session, bufferSize, maxIdleTime);
 	}
 }
